Negative SafPlatformAccessControl test
