+++ zowe-rest-api-commons-spring/src/test/java/org/zowe/commons/zos/security/platform/SafPlatformAccessControlTests.java
@@ -38,4 +38,11 @@ public class SafPlatformAccessControlTests {
     public void throwsExceptionOnInvalidLevel() {
         new MockPlatformAccessControl("test-saf-invalid.yml");
     }
+
+    @Test(expected = SafPlatformError.class)
+    public void returnSafPlatformErrorForInvalidClassNames() {
+        SafPlatformAccessControl badPlatformAccessControl = new SafPlatformAccessControl(
+                new BadMockPlatformClassFactory());
+        assertNull(badPlatformAccessControl.checkPermission("ZOWE", "SAMPLE.RESOURCE", AccessLevel.READ.getValue()));
+    }
 }
